requirements_sana.txt
absl-py==2.3.1
accelerate==1.10.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
args==0.1.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
braceexpand==0.1.7
came-pytorch==0.1.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clint==0.5.1
clip==0.2.0
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
datasets==4.2.0
deepspeed==0.18.0
diffusers==0.35.1
dill==0.4.0
DISTS-pytorch==0.1
einops==0.8.1
elatentlpips==0.2
exceptiongroup==1.3.0
fairscale==0.4.13
filelock==3.14.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.3.0
iopath==0.1.10
jmespath==0.10.0
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.15.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mmcv==2.2.0
mmdet==3.3.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
namex==0.1.0
networkx==3.3
ninja==1.13.0
numpy==2.1.2
omegaconf==2.3.0
open_clip_torch==3.2.0
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.3
peft==0.17.1
perlin_noise==1.14
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==13.4.2
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminaltables==3.1.10
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
tomli==2.3.0
torchao==0.10.0
torchmetrics==1.8.2
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.14
webdataset==1.0.2
Werkzeug==3.1.3
wrapt==1.17.3
xformers==0.0.29.post1
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
